% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{extract.file}
\alias{extract.file}
\title{Extract a specific file from a tar archive}
\usage{
extract.file(tar.file, extract.file, new.file, resultsdir)
}
\arguments{
\item{tar.file}{path to tar file}

\item{extract.file}{file to extract from tar archive}

\item{new.file}{output file name}

\item{resultsdir}{directory to extract files to}
}
\value{
none
}
\description{
Extract a specific file from a tar archive
}
