% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{extract.participant}
\alias{extract.participant}
\title{Extract participant ID from TCGA barcode}
\usage{
extract.participant(id)
}
\arguments{
\item{id}{TCGA barcode}
}
\value{
participant ID
}
\description{
Extract participant ID from TCGA barcode
}
