% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{extract.tissue}
\alias{extract.tissue}
\title{Extract tissue code from TCGA barcode}
\usage{
extract.tissue(id)
}
\arguments{
\item{id}{TCGA barcode}
}
\value{
tissue code
}
\description{
Extract tissue code from TCGA barcode
}
