% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{my.write.table}
\alias{my.write.table}
\title{Write table}
\usage{
my.write.table(x, filename)
}
\arguments{
\item{x}{data frame to write}

\item{filename}{output file name}
}
\value{
none
}
\description{
Write table
}
