% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{extract.tissue}
\alias{extract.tissue}
\title{Extract tissue code from TCGA barcode}
\usage{
extract.tissue(id)
}
\arguments{
\item{id}{TCGA barcode}
}
\value{
tissue code
}
\description{
The format of sample identifiers/barcodes is described here:
https://docs.gdc.cancer.gov/Encyclopedia/pages/TCGA_Barcode/
}
\details{
Here is a summary:
e.g. TCGA-3C-AAAU-01A-11D-A41Q-05
project TCGA
tissue source site 3C
participant AAAU
sample 01 (01-09 tumor, 10-19 normal, 20-29 controls)
vial A
portion 11
analyte D (as in DNA)
plate A41Q
analysis center 05

The following function extracts the participant identifier
from a sample id/barcode.
}
